api_gateway/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
httpx==0.26.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2 